adding new plugin class
