+++ hourglass/backends/sensu/cache.py
@@ -91,35 +91,5 @@ class SensuZoneMetadata(SensuBase, ZoneMetadata):
 
 class SensuZone(SensuBase, Zone):
 
-    models = [SensuCheck, SensuClient, SensuEvent, SensuStash, SensuResult]
-
-    def __init__(self, name, host, port, timeout):
-        self.name = name
-        self.host = host
-        self.port = port
-        self.timeout = timeout
-
-    @asyncio.coroutine
-    def query_api(self, session, url):
-        with aiohttp.Timeout(self.timeout):
-            response = yield from session.get(url)
-            return (yield from response.json())
-
-    @asyncio.coroutine
-    def update_objects(self, app, loop, model):
-        init_objects = []
-        try:
-            with aiohttp.ClientSession(loop=loop) as session:
-                url = "http://%s:%s/%s" % (self.host, self.port, model.uri)
-                app.logger.debug('Making request to %s' % url)
-                results = yield from self.query_api(session, url)
-        except aiohttp.errors.ClientError as e:
-            app.logger.error('Error updating %s cache for %s: %s' % (
-                model.__tablename__, self.name, e))
-            return init_objects
-        model.query.filter(model.zone_name == self.name).delete()
-        for result in results:
-            init_objects.append(model(self.name, result))
-        app.logger.info('Updated %s cache for %s' % (
-            model.__tablename__, self.name))
-        return init_objects
+    def __repr__(self):
+        return '<SensuZone %s>' % self.name
